Improved colonist hueristic/added #! and newlines
This commit introduced a better colonist hueristic based on a "priority"
dict. It's still pretty stupid, but less stupid than what came before.

A number of other files were modified in minor ways to make it more
convienient for me to work on them (because of weird ways I use
emacs). For example, main.py had a #!/usr/bin/python added to the
first line, and was made executable.

This commit is untested, and may not even run at all.